### 2020-08-20 #2589 @whedon
[PRE REVIEW]: rFBP: Replicated Focusing Belief Propagation algorithm
**Submitting author:** @Nico-Curti (<a href="http://orcid.org/0000-0001-5802-1195">Nico Curti</a>)
**Repository:** <a href="https://github.com/Nico-Curti/rFBP" target ="_blank">https://github.com/Nico-Curti/rFBP</a>
**Version:** v1.0.0
