### 2020-08-21 b56007f
[REVISION] move contribution footnote
