### 2020-08-24 #2589 @arokem
Hi @Nico-Curti and co-authors: thank you for this submission. I have now read the article and looked through the software. The next steps are for me to offer some comments on this initial submission and for us to find appropriate reviewers.

First, my comments on the article:

1. The readership of JOSS is expected to be very broad, so you will need to provide a bit more context for the software. For example, the first sentence of the article really drops the reader off the deep end, where more general context would have been helpful.

2. Since an important element in a JOSS paper is the potential research applications of the software, it would be good to flesh out the examples that you provided as bullet points. How was the software used in these applications? What kinds of research questions were answered? In general, as I understand it, since this is a re-implementation of an existing algorithm, what is the need that is addressed through this implementation of the algorithm?

Second: do you have particular suggestions for people who would be appropriate reviewers for your paper? If you include their GitHub handles here, please *do not* use the "@" sign before their name, so that I can take a look before tagging them into this issue.
