### 2020-08-30 #2589 @arokem
👋 @justusschock : would you be willing to review this article for JOSS? This submissions seems like a good match for your expertise.
