### 2020-08-30 #2589 @arokem
:wave: @georgiastuart : would you be willing to review this article for JOSS? It seems like a good match for your expertise.
