### 2020-09-01 #2589 @Nico-Curti
@arokem Thanks for your comments. We will provide an updated version of the paper with your suggestions in few days.
As for the second reviewer we would like to suggest professor [Brian Blais](https://github.com/bblais) since he is an expert in the field and a very competent Python programmer.
