### 2020-09-01 acd8dea
[REVISION] temporary fix the setuptools version
