### 2020-09-01 9bec66b
[REVISION] temporary fix the pip version just to be sure
