+++ .travis.yml
@@ -347,7 +347,11 @@ before_install:
 
   # install python packages
   - $PYTHON --version
+
+  # temporary solution for setuptools error (https://stackoverflow.com/questions/63663362/django-python3-on-install-i-get-parent-module-setuptools-not-loaded)
+  - if [[ "$WITH_PYTHON" == "ON" ]]; then $PYTHON -m pip install --user pip==20.2.1                                                           ;fi
   - if [[ "$WITH_PYTHON" == "ON" ]]; then $PYTHON -m pip install --user setuptools==49.3.0                                                    ;fi
+
   - if [[ "$WITH_PYTHON" == "ON" ]]; then $PYTHON -m pip install --user numpy==1.15                                                           ;fi
   - if [[ "$WITH_PYTHON" == "ON" ]]; then $PYTHON -m pip install --user cython==0.29                                                          ;fi
   - if [[ "$WITH_PYTHON" == "ON" ]]; then $PYTHON -m pip install --user scipy==1.2.1                                                          ;fi
