### 2020-09-01 #2589 @arokem
Hi @bblais -- would you be willing to review this software contribution to [The Journal of Open Source Software](https://joss.theoj.org/)? The reviews are conducted openly, through a GitHub issue. More instructions are available [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)
