### 2020-09-10 #2589 @arokem
:wave: @DanielLenz : would you be willing to review this article for JOSS?
