### 2020-09-10 #2589 @bblais
> Hi @bblais -- would you be willing to review this software contribution to [The Journal of Open Source Software](https://joss.theoj.org/)? The reviews are conducted openly, through a GitHub issue. More instructions are available [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)

Sure, I can review (didn't notice this invite the first time).  I see the editorial guide, but is there a guide for the process of submitting a review?  is it by individual issues or all at once?  is there a deadline or expected timeframe for review?  thanks!
