### 2020-09-10 #2589 @arokem
@bblais , @justusschock : please head over to #2663. There you will have a checklist with items to check off. As you go along, feel free to comment and make suggestions in that review issue, or to post issues on the software repository (might want to mention that the issue stems from the JOSS review). And please do let me know if you have any further questions.
