### 2020-09-19 #2663 @arokem
After a private exchange with @bblais we've ascertained that he has an apparent conflict-of-interest, due to a recent joint publication with one of the authors of this paper. That means that we will have to find another reviewer to replace him here. For now, I will edit the checklist at the top to indicate that @bblais is no longer a reviewer and unassign him. Thanks @bblais! I hope we find another opportunity to call on you to review for JOSS.
