### 2020-09-28 Nico-Curti/rFBP#2 @justusschock
LICENSE File
Hi,
first I want to thank you for this incredible work.

While reviewing for JOSS (https://github.com/openjournals/joss-reviews/issues/2663), I noted that your license file is markdown and not a plain text file as required :)
