### 2020-09-28 Nico-Curti/rFBP#3 @justusschock
Docstrings
During my review for JOSS (https://github.com/openjournals/joss-reviews/issues/2663), I also found that for your C++ Code there are not always docstrings for functions. Could you maybe add these and while on it make sure they can be parsed from oxygen or whatever?

Also your Python Docstrings could be cleaned up a bit to be a bit more consistent regarding the parameter types. If you revisit them, make sure they follow the sphinx format (or some other but sphinx is really easy to use). If you want to, you could also remove the parameter types and replace them by Python's type annotations.

This is basically a preparation for #4
