### 2020-09-28 Nico-Curti/rFBP#4 @justusschock
Functionality Documentation
Hi,

during my review for JOSS (https://github.com/openjournals/joss-reviews/issues/2663) I looked for what they call "functionality documentation". I was not able to find any API-Documentation on your GitHub page. Could you maybe add this?
