### 2020-09-28 #2663 @justusschock
@arokem I finished my review. Overall I really like the submission. Regarding the checklist above:
* So far there is no API Documentation / Only some docstrings
* The license file must be changed from Markdown to plain text

If the above mentioned points are finished, I don't see any blockers from my side.
