### 2020-10-03 Nico-Curti/rFBP#5 @DanielLenz
Install from PyPI: numpy not found
Hi! This issue is related to the JOSS review (https://github.com/openjournals/joss-reviews/issues/2663).

I tried to install the Python package from pip in a fresh venv and noticed that the requirements aren't installed as  they should be.

```bash

~ pipenv install ReplicatedFocusingBeliefPropagation

Installing ReplicatedFocusingBeliefPropagation…
Error:  An error occurred while installing ReplicatedFocusingBeliefPropagation!
Error text: Collecting replicatedfocusingbeliefpropagation
  Downloading ReplicatedFocusingBeliefPropagation-1.0.1.tar.gz (417 kB)

     command: /Users/lenzda/.local/share/virtualenvs/py-Ob9eXrK4/bin/python3.7m -c 'import sys, setuptools, tokenize; sys.argv[0] = '"'"'/private/var/folders/4x/6yrptfcn4kxdynxfy3qv1g20lmvv2v/T/pip-install-uu4qb3_7/replicatedfocusingbeliefpropagation/setup.py'"'"'; __file__='"'"'/private/var/folders/4x/6yrptfcn4kxdynxfy3qv1g20lmvv2v/T/pip-install-uu4qb3_7/replicatedfocusingbeliefpropagation/setup.py'"'"';f=getattr(tokenize, '"'"'open'"'"', open)(__file__);code=f.read().replace('"'"'\r\n'"'"', '"'"'\n'"'"');f.close();exec(compile(code, __file__, '"'"'exec'"'"'))' egg_info --egg-base /private/var/folders/4x/6yrptfcn4kxdynxfy3qv1g20lmvv2v/T/pip-pip-egg-info-vplpgjcz
         cwd: /private/var/folders/4x/6yrptfcn4kxdynxfy3qv1g20lmvv2v/T/pip-install-uu4qb3_7/replicatedfocusingbeliefpropagation/
    Complete output (5 lines):
    Traceback (most recent call last):
      File "<string>", line 1, in <module>
      File "/private/var/folders/4x/6yrptfcn4kxdynxfy3qv1g20lmvv2v/T/pip-install-uu4qb3_7/replicatedfocusingbeliefpropagation/setup.py", line 8, in <module>
        import numpy as np
    ModuleNotFoundError: No module named 'numpy'
ERROR: Command errored out with exit status 1: python setup.py egg_info Check the logs for full command output.

This is likely caused by a bug in ReplicatedFocusingBeliefPropagation. Report this to its maintainers.
✘ Installation Failed
```
