### 2020-10-03 Nico-Curti/rFBP#6 @DanielLenz
CONTRIBUTING.md doesn't exist
Hi! This issue is related to the JOSS review (https://github.com/openjournals/joss-reviews/issues/2663).

I noted that towards the end of the `README`, you refer to the `CONTRIBUTING.md` for instructions on how to contribute to this project, but this file doesn't exist.

The same is true for the issue template and for the PR template.
