### 2020-10-04 #2663 @DanielLenz
Hey @arokem , I've finished my review and agree with @justusschock in that I really like the submission and think the software is well-written and follows many best practices, big thanks for @Nico-Curti .

There are a couple of points that we raised on the projects github issue tracker which should be taken care of first. All of these are fairly minor and shouldn't take long.

Regarding the paper itself, there are a couple of minor points that I'll list here. All of these refer to the [latest draft compiled by wheadon](https://github.com/openjournals/joss-papers/blob/joss.02663/joss.02663/10.21105.joss.02663.pdf)

- The hyperlinks in the pdf didn't work for me. Is this normal or am I the only one experiencing this?
- Remove the line that contains the bibliography file name
- Affiliations: Please use one line per affiliation
- Can we move the "Joint first author" footnote closer to the author list, perhaps right next to the affiliations?
- First paragraph:
  - ~fast~ quickly converging
- Second paragraph:
  - ... has already been widely used...
- Third paragraph:
  - ... ill-posed conditions can be tackled
  - i.e. a prior known...
  - (_perfect learning_, Baldassi et al. ....)
- Fourth paragraph:
  - `seems most promising` isn't particularly specific - perhaps `... spin-glass models have emerged as the most promising ones`?
- Fifth paragraph
  - I don't quite understand the first reference: Zecchina et al. (Baldassi et al., 2016a). Zecchina isn't listed in the bibliography either.

Thanks for working on this project over the years and for submitting to JOSS, I'm really looking forward to seeing it published!

Could you ping us here once these minor issues are addressed? Thank you!
