### 2020-10-07 a97eaca
[REVISION] fix affiliation and solve JOSS paper issues
