+++ docs/authors.md
@@ -5,7 +5,7 @@
 <img align="right" width="180" height="183" src="https://avatars0.githubusercontent.com/u/24650975?s=400&v=4"> About me:
 - **Name**: Nico
 - **Surname**: Curti
-- **Profession**: PhD at Dept. of Experimental, Diagnostic and Specialty Medicine of Bologna University
+- **Profession**: PhD at eDIMESLab, Dept. of Experimental, Diagnostic and Specialty Medicine of Bologna University
 - **University**: University of Bologna
 - **Location**: Italy
 - **Web page**: [git](https://github.com/Nico-Curti), [unibo](https://www.unibo.it/sitoweb/nico.curti2)
@@ -27,7 +27,7 @@
 <img align="right" width="180" height="183" src="https://avatars2.githubusercontent.com/u/1419337?s=400&v=4"> About me:
 - **Name**: Enrico
 - **Surname**: Giampieri
-- **Profession**: PhD at Dept. of Experimental, Diagnostic and Specialty Medicine
+- **Profession**: PhD at eDIMESLab, Dept. of Experimental, Diagnostic and Specialty Medicine
 - **University**: University of Bologna
 - **Location**: Italy
 - **Web page**: [git](https://github.com/EnricoGiampieri), [unibo](https://www.unibo.it/sitoweb/enrico.giampieri)
