+++ AUTHORS.md
@@ -1,6 +1,6 @@
 # Authors #
 
 ----------
-- Nico Curti - Dept. of Experimental, Diagnostic and Specialty Medicine of Bologna University ([nico.curti2@unibo.it](mailto:nico.curti2@unibo.it))
+- Nico Curti - eDIMESLab, Dept. of Experimental, Diagnostic and Specialty Medicine of Bologna University ([nico.curti2@unibo.it](mailto:nico.curti2@unibo.it))
 - Daniele Dall'Olio - Department of Physics and Astronomy, University of Bologna ([daniele.dallolio@studio.unibo.it](mailto:daniele.dallolio@studio.unibo.it))
-- Enrico Giampieri - Dept. of Experimental, Diagnostic and Specialty Medicine of Bologna University ([enrico.giampieri@unibo.it](mailto:enrico.giampieri@unibo.it))
+- Enrico Giampieri - eDIMESLab, Dept. of Experimental, Diagnostic and Specialty Medicine of Bologna University ([enrico.giampieri@unibo.it](mailto:enrico.giampieri@unibo.it))
