### 2020-10-07 f5a9bda
[REVISION] add full description of the API in sphinx+doxygen
