+++ ReplicatedFocusingBeliefPropagation/rfbp/magnetization.py
@@ -23,10 +23,54 @@ def _check_mag (x, cls=BaseMag):
     raise ValueError('Incompatible type found. x must be a Mag')
 
 def lr (x):
+  '''
+  log1p for magnetizations
+
+  Parameters
+  ----------
+    x : float
+      Input variable
+
+  Returns
+  -------
+    res : float
+      value computed as log1p(exp(-2*abs(x)))
+
+  Example
+  -------
+  >>> import numpy as np
+  >>> from ReplicatedFocusingBeliefPropagation import magnetization as mag
+  >>> x = np.random.uniform(low=0., high=1.)
+  >>> assert mag.lr(x) >= 0.
+  '''
   return np.log1p(np.exp(-2 * np.abs(x)))
 
 @singledispatch
 def sign0 (x):
+  '''
+  Sign operation valid also for magnetizations
+
+  Parameters
+  ----------
+    x : float
+      Input variable
+
+  Returns
+  -------
+    res : float
+      sign evaluated as 1 - 2*signbit(x)
+
+  Example
+  -------
+  >>> import numpy as np
+  >>> from ReplicatedFocusingBeliefPropagation import MagP64
+  >>> from ReplicatedFocusingBeliefPropagation import magnetization as mag
+  >>> x = np.random.uniform(low=0., high=1.)
+  >>> m = MagP64(x)
+  >>> assert mag.sign0(x) in (-1, 1)
+  >>> assert mag.sign0(-x) == 1
+  >>> assert mag.sign0(m) in (-1, 1)
+  '''
   return 1 if np.sign(x) else -1
 
 @sign0.register(BaseMag)
@@ -34,6 +78,37 @@ def _ (x):
   return 1 if np.sign(x.mag) else -1
 
 def zeros (x):
+  '''
+  Fill array of magnetizations with zeros
+
+  Parameters
+  ----------
+    x : array-like
+      Input array or list of Mag objects
+
+  Returns
+  -------
+    None
+
+  Example
+  -------
+  >>> import numpy as np
+  >>> from ReplicatedFocusingBeliefPropagation import MagP64
+  >>> from ReplicatedFocusingBeliefPropagation import magnetization as mag
+  >>> x = np.random.uniform(low=0., high=1.)
+  >>> mags = [MagP64(_) for _ in range(10)]
+  >>> mag.zeros(mags)
+  >>> assert all((i.mag == 0 for i in mags))
+  >>>
+  >>> l = 3.14
+  >>> mag.zeros(3.14)
+    ValueError('zeros takes an iterable object in input')
+  >>>
+  >>> l = [MagP64(3.14), x]
+  >>> mag.zeros(l)
+    ValueError('Incompatible type found. x must be an iterable of Mags')
+  '''
+
   if not isinstance(x, Iterable):
     raise ValueError('zeros takes an iterable object in input')
 
@@ -45,6 +120,32 @@ def zeros (x):
 
 @singledispatch
 def zero (x):
+  '''
+  Set magnetization to zero
+
+  Parameters
+  ----------
+    x : Mag object
+      Input variable
+
+  Returns
+  -------
+    None
+
+  Example
+  -------
+  >>> import numpy as np
+  >>> from ReplicatedFocusingBeliefPropagation import MagP64
+  >>> from ReplicatedFocusingBeliefPropagation import magnetization as mag
+  >>> x = np.random.uniform(low=0., high=1.)
+  >>> m = MagP64(3.14)
+  >>> mag.zero(m)
+  >>>
+  >>> assert m.mag == 0.
+  >>> assert m.value == 0.
+  >>> mag.zero(x)
[…]
