+++ ReplicatedFocusingBeliefPropagation/rfbp/MagP64.py
@@ -16,12 +16,21 @@ class MagP64 (BaseMag):
 
   def __init__ (self, x):
     '''
-    MagP64 specialization
+    MagP64 specialization.
+    The `MagP64` type allows fast executions with inexact outcomes by neglecting all `tanh` operations.
 
     Parameters
     ----------
       x : float
         Value to magnetify
+
+    Example
+    -------
+    >>> import numpy as np
+    >>> from ReplicatedFocusingBeliefPropagation.rfbp.Mag import MagP64
+    >>>
+    >>> x = np.random.uniform(low=0., high=1)
+    >>> m = MagP64(x)
     '''
     super(MagP64, self).__init__(x)
 
@@ -29,6 +38,19 @@ class MagP64 (BaseMag):
   def value (self):
     '''
     Return the mag value
+
+    Returns
+    -------
+      x : float
+        In MagP64 the value is equal to the magnetization since the tanh operation is neglected
+
+    Example
+    -------
+    >>> from ReplicatedFocusingBeliefPropagation import MagP64
+    >>> x = np.random.uniform(low=0, high=1.)
+    >>> m = MagP64(x)
+    >>> assert np.isclose(m.mag, x)
+    >>> assert np.isclose(m.value, x)
     '''
     return self.mag
 
@@ -36,6 +58,18 @@ class MagP64 (BaseMag):
   def magformat (self):
     '''
     Return the mag description
+
+    Returns
+    -------
+      plain : str
+        The MagP64 type corresponds to a plain operation
+
+    Example
+    -------
+    >>> from ReplicatedFocusingBeliefPropagation import MagP64
+    >>> m = MagP64(3.14)
+    >>> m.magformat
+      'plain'
     '''
     return 'plain'
 
@@ -43,6 +77,27 @@ class MagP64 (BaseMag):
   def convert (x):
     '''
     Convert a float to a mag value (as a constructor)
+
+    Parameters
+    ----------
+      x : float
+        The number to convert
+
+    Returns
+    -------
+      m : MagP64
+        Convert any-number to a MagP64 type
+
+    Example
+    -------
+    >>> import numpy as np
+    >>> from ReplicatedFocusingBeliefPropagation import MagP64
+    >>>
+    >>> x = np.random.uniform(low=0., high=10)
+    >>> m1 = MagP64.convert(x)
+    >>> m2 = MagP64(x)
+    >>> assert m1.mag == m2.mag
+    >>> assert m1.value == m2.value
     '''
     return MagP64(x)
 
@@ -50,6 +105,30 @@ class MagP64 (BaseMag):
   def couple (x1, x2):
     '''
     Combine two mags as diff / sum
+
+    Parameters
+    ----------
+      x1 : float
+        The first element of the operation
+      x2 : float
+        The second element of the operation
+
+    Returns
+    -------
+      x : float
+        In MagP64 the value is equal to the magnetization since the tanh operation is neglected
+
+    Example
+    -------
+    >>> import numpy as np
+    >>> from ReplicatedFocusingBeliefPropagation import MagP64
+    >>>
+    >>> x = np.random.uniform(low=0., high=10)
+    >>> y = np.random.uniform(low=0., high=10)
+    >>> mx = MagP64.couple(x, y)
+    >>> my = MagP64.couple(y, x)
+    >>> assert np.isclose(abs(mx.mag), abs(my.mag))
+    >>> assert np.isclose(abs(mx.value), abs(my.value))
     '''
     return MagP64( (x1 - x2)/(x1 + x2) )
 
[…]
