+++ include/cmd.h
@@ -8,7 +8,11 @@
 #include <string>
 
 /**
-* @brief Parse command line for training procedure using parse_args library
+* @brief Parse command line for training procedure using parse_args library.
+*
+* @note This function is just an utility for the usage of the library from command line interface.
+* It can be substituted by any equivalent command-line parser library.
+* The variable are set into the function and thus are all passed by reference.
 *
 * @param[in] argc number of arguments in command line
 * @param[in] argv list of arguments in command line
@@ -37,7 +41,11 @@
 void parse_training_fbp (int argc, char *argv[], std :: string & patternsfile, std :: string & output, bool & bin, std :: string & del, long int & K, long int & max_iters, long int & seed, double & randfact, double & damping, std :: string & accuracy1, std :: string & accuracy2, std :: string & fprotocol, double & epsil, int & nth, long int & max_steps, int & mag, std :: string & inmess, std :: string & outmess, std :: string & delmess, bool & binmess);
 
 /**
-* @brief Parse command line for testing procedure using parse_args library
+* @brief Parse command line for testing procedure using parse_args library.
+*
+* @note This function is just an utility for the usage of the library from command line interface.
+* It can be substituted by any equivalent command-line parser library.
+* The variable are set into the function and thus are all passed by reference.
 *
 * @param[in] argc number of arguments in command line
 * @param[in] argv list of arguments in command line

[…]
