### 2020-10-07 1d4cf72
[REVISION] add setup warning for setuptools issue
