+++ appveyor.yml
@@ -1,24 +1,97 @@
-image: Visual Studio 2017
 clone_folder: c:\projects\rFBP
-configuration: Release
+configuration: Debug
+
+image:
+  - Visual Studio 2017
+  - Visual Studio 2019
 
 environment:
-    WORKSPACE: C:\projects
-    matrix:
-    - platform: Win32
-      COMPILER: vs
-      MSVC_SETUP_ARG: x86
-      PYTHON: "C:\\Python36"
-      PYTHON_VERSION: "3.6.3"
-      PYTHON_ARCH: "32"
-      MINICONDA: "C:\\Miniconda36"
-    - platform: Win64
-      COMPILER: vs
-      MSVC_SETUP_ARG: x64
-      PYTHON: "C:\\Python36-x64"
-      PYTHON_VERSION: "3.6.3"
-      PYTHON_ARCH: "64"
-      MINICONDA: "C:\\Miniconda36-x64"
+
+  WORKSPACE: C:\projects
+  # MSVC v2017
+  matrix:
+  - platform: Win32
+    APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2017
+    PYTHON: "C:\\Python36"
+    PYTHON_VERSION: "3.6.3"
+    PYTHON_ARCH: "32"
+    MINICONDA: "C:\\Miniconda36"
+    COMPILER: vs
+    MSVC_SETUP_PATH: C:\Program Files (x86)\Microsoft Visual Studio\2017\Community\VC\Auxiliary\Build\vcvarsall.bat
+    MSVC_SETUP_ARG: x86
+  - platform: Win32
+    APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2017
+    PYTHON: "C:\\Python36"
+    PYTHON_VERSION: "3.6.3"
+    PYTHON_ARCH: "32"
+    MINICONDA: "C:\\Miniconda36"
+    COMPILER: vs
+    MSVC_SETUP_PATH: C:\Program Files (x86)\Microsoft Visual Studio\2017\Community\VC\Auxiliary\Build\vcvarsall.bat
+    MSVC_SETUP_ARG: x86
+  - platform: Win64
+    APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2017
+    PYTHON: "C:\\Python36-x64"
+    PYTHON_VERSION: "3.6.3"
+    PYTHON_ARCH: "64"
+    MINICONDA: "C:\\Miniconda36-x64"
+    COMPILER: vs
+    MSVC_SETUP_PATH: C:\Program Files (x86)\Microsoft Visual Studio\2017\Community\VC\Auxiliary\Build\vcvarsall.bat
+    MSVC_SETUP_ARG: x64
+  - platform: Win64
+    APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2017
+    PYTHON: "C:\\Python36-x64"
+    PYTHON_VERSION: "3.6.3"
+    PYTHON_ARCH: "64"
+    MINICONDA: "C:\\Miniconda36-x64"
+    COMPILER: vs
+    MSVC_SETUP_PATH: C:\Program Files (x86)\Microsoft Visual Studio\2017\Community\VC\Auxiliary\Build\vcvarsall.bat
+    MSVC_SETUP_ARG: x64
+
+  # MSVC v2019
+  - platform: Win32
+    APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2019
+    PYTHON: "C:\\Python36"
+    PYTHON_VERSION: "3.6.3"
+    PYTHON_ARCH: "32"
+    MINICONDA: "C:\\Miniconda36"
+    COMPILER: vs
[…]
