### 2020-10-07 f5f0f26
[REVISION] fix the requirements for the docs
