### 2020-10-07 f3d0384
[REVISION] remove MagT64 testing for log1pxy
