### 2020-10-07 08d45e5
[REVISION] fix docs dependencies
