+++ docs/sphinxtext/source/conf.py
@@ -24,6 +24,7 @@ author = "Nico Curti, Daniele Dall'Olio, Enrico Giampieri"
 # The full version, including alpha/beta/rc tags
 release = '1.0.1'
 
+master_doc = 'index'
 
 # -- General configuration ---------------------------------------------------
 
