### 2020-10-07 3661efa
[REVISION] fix requirements for docs
