+++ docs/sphinxtext/source/conf.py
@@ -11,9 +11,49 @@
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 #
 import os
+import re
 import sys
+
 sys.path.insert(0, os.path.abspath('../../../ReplicatedFocusingBeliefPropagation/'))
 
+def read_version (CMakeLists):
+  """
+  Read version from variables set in CMake file
+
+  Parameters
+  ----------
+  CMakeLists : CMake file path
+
+  Returns
+  -------
+  version : tuple
+    Version as (major, minor, revision)
+  """
+  major = re.compile(r'set\s+\(RFBP_MAJOR\s+(\d+)\)')
+  minor = re.compile(r'set\s+\(RFBP_MINOR\s+(\d+)\)')
+  revision = re.compile(r'set\s+\(RFBP_REVISION\s+(\d+)\)')
+
+  with open(CMakeLists, 'r') as fp:
+    cmake = fp.read()
+
+  major_v = major.findall(cmake)[0]
+  minor_v = minor.findall(cmake)[0]
+  revision_v = revision.findall(cmake)[0]
+
+  version = map(int, (major_v, minor_v, revision_v))
+
+  return tuple(version)
+
+try:
+
+  LOCAL = os.path.dirname(__file__)
+
+except NameError:
+
+  LOCAL = ''
+
+VERSION = read_version(os.path.join(LOCAL, '..', '..', '..', 'CMakeLists.txt'))
+__version__ = '.'.join(map(str, VERSION))
 
 # -- Project information -----------------------------------------------------
 
@@ -22,7 +62,7 @@ copyright = "2020, Nico Curti, Daniele Dall'Olio, Enrico Giampieri"
 author = "Nico Curti, Daniele Dall'Olio, Enrico Giampieri"
 
 # The full version, including alpha/beta/rc tags
-release = '1.0.1'
+release = __version__
 
 master_doc = 'index'
 
