+++ docs/requirements.txt
@@ -1,5 +1,11 @@
+numpy>=1.15
+cython>=0.29
+scipy>=1.2.1
+scikit-learn>=0.20.3
+requests>=2.22.0
+
 breathe<4.13.0
-sphinx
+sphinx==1.3.1
 sphinx-rtd-theme
 sphinxcontrib-napoleon
 sphinxcontrib-programoutput
