+++ docs/sphinxtext/source/conf.py
@@ -71,7 +71,8 @@ master_doc = 'index'
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
-extensions = ['sphinx.ext.napoleon',
+extensions = ['sphinx.ext.autodoc',
+              'sphinx.ext.napoleon',
 						  'sphinx.ext.viewcode',
 						  'breathe'
 ]
