+++ docs/requirements.txt
@@ -5,7 +5,7 @@ scikit-learn>=0.20.3
 requests>=2.22.0
 
 breathe<4.13.0
-sphinx==1.3.1
+sphinx==1.8
 sphinx-rtd-theme
 sphinxcontrib-napoleon
 sphinxcontrib-programoutput
