### 2020-10-07 bef9e6b
[REVISION] fix manifest
