+++ .github/workflows/docs.yml
@@ -32,7 +32,11 @@ jobs:
         python -m pip install --upgrade pip
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
 
-    - name: Install dependencies
+    - name: Install rfbp
+      run: |
+        python setup.py develop --user
+
+    - name: Install docs dependencies
       run: |
         python -m pip install -r docs/requirements.txt
 
