+++ setup.py
@@ -289,7 +289,8 @@ setup(
                                                          'ReplicatedFocusingBeliefPropagation.*'],
                                                 exclude=('test', 'example')),
   include_package_data          = True,
-  data_files                    = [('', ['CMakeLists.txt', 'README.md', 'LICENSE'])],
+  data_files                    = [('', ['CMakeLists.txt', 'README.md', 'LICENSE']),
+                                   ('scripts', [os.path.join(here, 'scripts', 'download_atanherf.py')])],
   platforms                     = 'any',
   classifiers                   = [
                                     # Full list: https://pypi.python.org/pypi?%3Aaction=list_classifiers
