+++ setup.py
@@ -290,7 +290,7 @@ setup(
                                                 exclude=('test', 'example')),
   include_package_data          = True,
   data_files                    = [('', ['CMakeLists.txt', 'README.md', 'LICENSE']),
-                                   ('scripts', [os.path.join(here, 'scripts', 'download_atanherf.py')])],
+                                   ('scripts', ['./scripts/download_atanherf.py'])],
   platforms                     = 'any',
   classifiers                   = [
                                     # Full list: https://pypi.python.org/pypi?%3Aaction=list_classifiers
