### 2020-10-07 Nico-Curti/rFBP#2 @Nico-Curti
Hi, @justusschock. Thanks for your comments.
We followed your request and we converted the LICENSE file from Markdown to plain-text.
You can find the reviewed version of the file [here](https://github.com/Nico-Curti/rFBP/blob/master/LICENSE).
Let us know if there are other issues.
