### 2020-10-07 Nico-Curti/rFBP#3 @Nico-Curti
Hi @justusschock. Thanks for your comments.
We have improved the documentation following your comments. In the [last version](https://github.com/Nico-Curti/rFBP/commit/aeb2fef1d5d2bdb34cb53c3b4fc9f94cff2ffbe5) of the code we hope to have fixed all the Python docstrings and we have improved the description of the C++ APIs.
- We used sphinx for the Python documentation and doxygen for the C++ counterpart as you have suggested.
- We have also included a [small CI](https://github.com/Nico-Curti/rFBP/actions?query=workflow%3A%22rFBP+Docs+CI%22) for the documentation to check the build at every commit.
- For the parameter type we used the numpydoc [standard](https://numpydoc.readthedocs.io/en/latest/format.html).

You can find the final result of our documentation at the link: [docs](https://rfbp.readthedocs.io/en/latest/?badge=latest)

Let us know if you have any other comment or request about this topic.
