### 2020-10-07 fc8ae71
[REVISION] add warnings about pip installation
