+++ README.md
@@ -186,6 +186,8 @@ The easiest way to install the package is to use `pip`
 python -m pip install ReplicatedFocusingBeliefPropagation
 ```
 
+> :warning: The setup file requires the `Cython` and `Numpy` packages, thus make sure to pre-install them!
+
 The `Python` installation can be performed with or without the `C++` installation.
 The `Python` installation is always executed using [`setup.py`](https://github.com/Nico-Curti/blob/master/setup.py) script.
 
