### 2020-10-07 Nico-Curti/rFBP#4 @Nico-Curti
Hi,
The topic of this issue is very closed to #3 , thus we could close this issue and merge the comments in #3.
