### 2020-10-07 Nico-Curti/rFBP#6 @Nico-Curti
Hi @DanielLenz . Thanks for your comment.
We moved all the template files related to PR, issues and contributions without fixing the hyperlink.
In the latest version of the code we hope to have fixed all the references.
If you find other broken hyperlinks, please let us know.
