### 2020-10-07 Nico-Curti/rFBP#5 @Nico-Curti
Hi @DanielLenz. Thanks for you comment.
The [numpy](https://github.com/Nico-Curti/rFBP/blob/fc8ae71a8ff58858f6800eeb3a3f25a56c143d18/setup.py#L9) and [cython](https://github.com/Nico-Curti/rFBP/blob/fc8ae71a8ff58858f6800eeb3a3f25a56c143d18/setup.py#L31) packages are required for the C++ compilation and thus they need to be installed beforehand.
We couldn't find any way to make pip install them before running the setup.py
Therefore we put the warning for the user in the [documentation](https://github.com/Nico-Curti/rFBP/blob/fc8ae71a8ff58858f6800eeb3a3f25a56c143d18/README.md#L189)  asking to install numpy and cython before running the pip command.
The same issue is not verified in the travis CI since we directly call the setup from the cloned repository.

If you have any suggestion about how overcome this issue please let us know.
