### 2020-10-07 #2663 @Nico-Curti
Hi @DanielLenz
We worked on your issues, trying to address all of your points.

> One point that I ran into, but didn't raise an issue about is the loading of the C++ library on Mac OS with clang to run the tests. You mention this in your travis config and I can confirm the issue, but I'm unsure how to address it.

We don't have a MacOS machine available in any sort and we aren't experts on this system. We have tried to solve the issue using the travis CI but unfortunately we haven't found any patch. If someone else have any suggestion please let us know.

> The hyperlinks in the pdf didn't work for me. Is this normal or am I the only one experiencing this?

The hyperlinks don't work on the online version of the paper as displayed by github. They seems to work fine if you download the pdf.

> Remove the line that contains the bibliography file name

We have removed the duplicated lines from the paper as you have pointed out. Let us know if the issue persists.

> Affiliations: Please use one line per affiliation

We follow the guidelines on the JOSS website but the affiliations are still displayed on the same line. Maybe @arokem could suggest us how we can fix it.

> Can we move the "Joint first author" footnote closer to the author list, perhaps right next to the affiliations?

We noticed that other already published papers in JOSS has the footnote at the end of the page. Also in this case we can fix this issue but we need some help by @arokem

About the typos that you have highlighted we have followed all your comments and fix the paper document.
