### 2020-10-09 Nico-Curti/rFBP#5 @DanielLenz
There's a couple of ways to address this I think, some [are described here](https://luminousmen.com/post/resolve-cython-and-numpy-dependencies).

I'll go ahead and test this on my machine as soon as possible - I can highly relate to the frustration of trying to debug mac OS issues via travis, it just takes forever and you run into the most obscure errors...
