### 2020-10-11 #2663 @DanielLenz
Hi @arokem , @Nico-Curti ,

I've taken another detailed look at the paper and the repo, and I'm happy to say that I recommend the publication at this point.

There are some minor formatting issues that we discussed, perhaps @arokem can shed some light on this.

Thanks again for your submission and for including the feedback so quickly!
