### 2020-10-11 #2663 @arokem
Thanks @DanielLenz! I will look into the formatting issues.

@justusschock : when you get a chance, could you please take a look and see whether the outstanding issues you had are addressed in the revision? In particular, I see that you have two check-boxes un-checked at this point. Are both of these points (license and documentation of functionality) addressed?
