### 2020-10-12 377cea0
[REVISION] fix setup requirements and improve version
