+++ README.md
@@ -187,6 +187,7 @@ python -m pip install ReplicatedFocusingBeliefPropagation
 ```
 
 > :warning: The setup file requires the `Cython` and `Numpy` packages, thus make sure to pre-install them!
+> We are working on some workarounds to solve this issue.
 
 The `Python` installation can be performed with or without the `C++` installation.
 The `Python` installation is always executed using [`setup.py`](https://github.com/Nico-Curti/blob/master/setup.py) script.
