+++ docs/sphinxtext/source/Python.rst
@@ -12,6 +12,7 @@ The easiest way to install the package is using `pip`
 .. warning::
 
 	The setup file requires the `Cython` and `Numpy` packages, thus make sure to pre-install them!
+	We are working on some workarounds to solve this issue.
 
 The `Python` installation can be performed with or without the `C++` installation.
 The `Python` installation is always executed using `setup.py`_ script.
