+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.2] - 2020-10-12
+
+### Added
+
+- Doxygen documentation of the C++ APIs
+- Sphinx documentation of the Python APIs
+- CI for the documentation build
+
+### Fixed
+
+- python setup using pip, fixing dependencies and installation paths
+
+
 ## [1.0.0] - 2020-07-03
 
 ### Added
