### 2020-10-16 d80a555
[REVISION] add BUILD_DOCS option in CMake
