+++ docs/Doxyfile.in
@@ -1,6 +1,6 @@
 DOXYFILE_ENCODING = UTF-8
 PROJECT_NAME = rFBP
-OUTPUT_DIRECTORY = ./docs/sphinxtext/source/
+OUTPUT_DIRECTORY = @CMAKE_SOURCE_DIR@/docs/sphinxtext/source/
 CREATE_SUBDIRS = NO
 OUTPUT_LANGUAGE = English
 ALWAYS_DETAILED_SEC = NO
@@ -22,6 +22,6 @@ XML_OUTPUT = doxydoc
 # RTF_HYPERLINKS = NO
 BUILTIN_STL_SUPPORT = YES
 EXTRACT_ALL = YES
-INPUT = ./include/
+INPUT = @CMAKE_CURRENT_SOURCE_DIR@/../include/
 RECURSIVE = YES
 HIDE_UNDOC_RELATIONS = YES
