### 2020-10-16 a827966
[REVISION] fix paper reference and update xml docs
