+++ paper/paper.bib
@@ -67,7 +67,11 @@
 @inproceedings{Krauth1989StorageCO,
   title={Storage capacity of memory networks with binary couplings},
   author={Werner Krauth and Marc M{\'e}zard},
-  year={1989}
+  year={1989},
+  journal={Journal de Physique},
+  volume={50},
+  pages={3057-3066},
+  doi={10.1051/jphys:0198900500200305700}
 }
 
 @article{Huang_2014,
