+++ paper/paper.md
@@ -10,10 +10,10 @@ tags:
 - python3
 - cpp17
 authors:
-- name: Nico Curti
+- name: Nico Curti^[co-first author]
   orcid: 0000-0001-5802-1195
   affiliation: 1
-- name: Daniele Dall'Olio^[Joint first author]
+- name: Daniele Dall'Olio^[co-first author]
   orcid: 0000-0003-0196-6870
   affiliation: 3
 - name: Daniel Remondini
