### 2020-10-16 #2663 @Nico-Curti
Hi @arokem,
I have solved both the issues.
Let me know if everything is ok now.
