### 2020-10-16 #2663 @arokem
Thank you. Could you please make sure that the Krauth and Mézard reference has the same format as the other references? It currently has the word "In" before the name of the journal (not needed) and the word "Vol" explicitly mentioned. It should look like this:

Krauth, W., & Mézard, M. (1989). Storage capacity of memory networks with binary couplings. *Journal de Physique*, 50, 3057–3066. doi:10.1051/jphys: 0198900500200305700
