+++ paper/paper.bib
@@ -64,11 +64,11 @@
   year={1992}
 }
 
-@inproceedings{Krauth1989StorageCO,
+@article{Krauth1989StorageCO,
   title={Storage capacity of memory networks with binary couplings},
-  author={Werner Krauth and Marc M{\'e}zard},
+  author={W. Krauth and M. M{\'e}zard},
+  journal={Journal De Physique},
   year={1989},
-  journal={Journal de Physique},
   volume={50},
   pages={3057-3066},
   doi={10.1051/jphys:0198900500200305700}
