+++ pyproject.toml
@@ -1,0 +1,3 @@
+[build-system]
+requires = ["setuptools>=18.0", "wheel", "numpy>=1.15", "Cython>=0.29"]
+build-backend = "setuptools.build_meta"
