### 2020-10-19 #2663 @Nico-Curti
Hi @arokem.
We have just uploaded the latest version of the package as [v1.0.3](https://github.com/Nico-Curti/rFBP/releases/tag/v1.0.3). The corresponding Zenodo doi is [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4106175.svg)](https://doi.org/10.5281/zenodo.4106175).
Let me know if everything is ok now.
