### 2020-10-19 Nico-Curti/rFBP#3 @Nico-Curti
I close the issue since the problem seems to be solved. If there are any other comments on this topic just let me know.
