### 2020-10-19 Nico-Curti/rFBP#5 @Nico-Curti
Hi @DanielLenz,
We have found a possible solution for our installation issue [here](https://snarky.ca/what-the-heck-is-pyproject-toml/). Using a [pyproject.toml](https://github.com/Nico-Curti/rFBP/blob/master/pyproject.toml) file the pip installation should pre-install all the requirements (cython and numpy in our specific case). In the [latest release](https://github.com/Nico-Curti/rFBP/releases/tag/v1.0.3) we have added this file (and some comments in the documentation to warning the users). We hope we have fixed the issue in this way.
