### 2020-10-20 #2663 @kyleniemeyer
Hi @Nico-Curti, I'm the EIC on duty this week, and doing some final checks before publishing your article. Can you fix a few small things?

- [ ] The paper is missing a Statement of Need section, which is a required element (as per [our docs](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain)); not sure how this didn't come up before, since whedon looks for this at submission, but it might have been because of the original compilation problems.
- [ ] In the third paragraph, please put a comma after "i.e"
- [ ] On the second page, this is a bit confusing: "...despite it is a quite efficient implementation the Julia programming language stays on difficult and far from many users." Can you rewrite this bit?
- [ ] In the same paragraph, I see "... C++ library called Scorer for further details)", which seems to be missing the starting parenthesis.
