### 2020-10-20 #2663 @Nico-Curti
Hi @kyleniemeyer,
We have just uploaded a reviewed version of the paper following your suggestions.
Please let me know if there are other issues.
